feat: feat/restructure-ansible
